requirements.txt
python-telegram-bot==21.4
httpx==0.27.0
python-dotenv==1.0.1

